+++ js/map.js
@@ -1,79 +1,73 @@
-import { offerForm } from './form.js';
+import { offers } from './data.js';
+import { adForm, filters, setAddressField } from './form.js';
 import { createCard } from './offer-card.js';
+import { enableForm } from './utile.js';
 
+const COORDINATES = {
+  lat: 35.68421,
+  lng: 139.75107,
+};
 const map = L.map('map-canvas');
-
 const markerGroup = L.layerGroup().addTo(map);
-
-const addressField = offerForm.querySelector('#address');
-
 const mainPinIcon = L.icon({
   iconUrl: 'img/main-pin.svg',
   iconSize: [52, 52],
   iconAnchor: [25,52],
 });
-
 const pinIcon = L.icon({
   iconUrl: 'img/pin.svg',
   iconSize: [40, 40],
   iconAnchor: [20,40],
 });
-
 const mainPinMarker = L.marker(
   {
     lat: 0,
     lng: 0,
   },
   {
-    draggable:true,
+    draggable: true,
     icon: mainPinIcon,
   }
 );
 
-export const mapInit = (coordinates) => {
-  map.setView(coordinates, 13);
+const addPinMarkers = (ads) => {
+  markerGroup.clearLayers();
 
-  L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
-    maxZoom: 19,
-    attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>'
-  }).addTo(map);
-
-  mainPinMarker.setLatLng(coordinates);
-  mainPinMarker.addTo(map);
-
-  mainPinMarker.on('moveend', (evt) => {
-
-    const {lat, lng} = evt.target.getLatLng();
-
-    addressField.value = `${lat.toFixed(5)}, ${lng.toFixed(5)}`;
-  });
-};
-
-const createOfferPinMarkers = (offers) => {
-  offers.forEach((offer) => {
-    const {location:{lat,lng}} = offer;
+  ads.forEach((ad) => {
+    const {location:{lat,lng}} = ad;
 
     const marker = L.marker(
       {
         lat,
         lng
       },
-
       {
         icon: pinIcon
       }
     );
 
     marker.addTo(markerGroup)
-      .bindPopup(createCard(offer));
+      .bindPopup(createCard(ad));
   });
 };
 
-export const setOfferPins = (offers) => {
-  markerGroup.clearLayers();
-  createOfferPinMarkers(offers);
-};
+map.on('load', () => {
+  enableForm(adForm);
+  enableForm(filters);
+  addPinMarkers(offers);
+});
 
-export const setOnMapLoad = (fn) => {
-  map.on('load', fn);
-};
+map.setView(COORDINATES, 13);
+
+L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
+  maxZoom: 19,
+  attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>'
+}).addTo(map);
+
+mainPinMarker.setLatLng(COORDINATES);
+mainPinMarker.addTo(map);
+mainPinMarker.on('moveend', (evt) => {
+  const {lat, lng} = evt.target.getLatLng();
+
+  setAddressField(lat, lng);
+});
